REPL-139 Moved classes in proper packages and re-define the exception class hierarchy.
