+++ replication-api-impl/src/main/java/com/connexta/replication/api/impl/data/FilterManagerImpl.java
@@ -13,11 +13,11 @@
  */
 package com.connexta.replication.api.impl.data;
 
-import com.connexta.ion.replication.api.NonTransientReplicationPersistenceException;
-import com.connexta.ion.replication.api.NotFoundException;
-import com.connexta.ion.replication.api.RecoverableReplicationPersistenceException;
-import com.connexta.ion.replication.api.TransientReplicationPersistenceException;
 import com.connexta.replication.api.data.Filter;
+import com.connexta.replication.api.data.NonTransientReplicationPersistenceException;
+import com.connexta.replication.api.data.NotFoundException;
+import com.connexta.replication.api.data.RecoverableReplicationPersistenceException;
+import com.connexta.replication.api.data.TransientReplicationPersistenceException;
 import com.connexta.replication.api.impl.persistence.pojo.FilterPojo;
 import com.connexta.replication.api.impl.persistence.spring.FilterRepository;
 import com.connexta.replication.api.persistence.FilterManager;
