+++ replication-api-impl/src/main/java/com/connexta/replication/api/impl/data/ReplicationItemImpl.java
@@ -15,8 +15,8 @@ package com.connexta.replication.api.impl.data;
 
 import com.connexta.ion.replication.api.Action;
 import com.connexta.ion.replication.api.Status;
-import com.connexta.ion.replication.api.UnsupportedVersionException;
 import com.connexta.replication.api.data.ReplicationItem;
+import com.connexta.replication.api.data.UnsupportedVersionException;
 import com.connexta.replication.api.impl.persistence.pojo.ItemPojo;
 import com.google.common.annotations.VisibleForTesting;
 import java.util.Date;
