+++ replication-api-impl/src/main/java/com/connexta/replication/api/impl/data/SiteImpl.java
@@ -13,11 +13,11 @@
  */
 package com.connexta.replication.api.impl.data;
 
-import com.connexta.ion.replication.api.InvalidFieldException;
-import com.connexta.ion.replication.api.UnsupportedVersionException;
+import com.connexta.replication.api.data.InvalidFieldException;
 import com.connexta.replication.api.data.Site;
 import com.connexta.replication.api.data.SiteKind;
 import com.connexta.replication.api.data.SiteType;
+import com.connexta.replication.api.data.UnsupportedVersionException;
 import com.connexta.replication.api.impl.persistence.pojo.SitePojo;
 import com.google.common.annotations.VisibleForTesting;
 import java.net.MalformedURLException;
