+++ replication-api/src/main/java/com/connexta/replication/api/data/NotFoundException.java
@@ -11,23 +11,17 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package com.connexta.ion.replication.api;
+package com.connexta.replication.api.data;
 
-/**
- * Thrown when a pojo cannot be deserialized or an object cannot be serialized because of an invalid
- * field.
- */
+/** Persistence exception indicating when a data entry that should be stored couldn't be found. */
 @SuppressWarnings("squid:MaximumInheritanceDepth" /* Exception class hierarchy */)
-public class InvalidFieldException extends NonTransientReplicationPersistenceException {
-  /** Instantiates a new exception. */
-  public InvalidFieldException() {}
-
+public class NotFoundException extends NonTransientReplicationPersistenceException {
   /**
    * Instantiates a new exception.
    *
    * @param message the message for the exception
    */
-  public InvalidFieldException(String message) {
+  public NotFoundException(String message) {
     super(message);
   }
 
@@ -37,7 +31,7 @@ public class InvalidFieldException extends NonTransientReplicationPersistenceExc
    * @param message the message for the exception
    * @param cause the cause for the exception
    */
-  public InvalidFieldException(String message, Throwable cause) {
+  public NotFoundException(String message, Throwable cause) {
     super(message, cause);
   }
 
@@ -46,7 +40,7 @@ public class InvalidFieldException extends NonTransientReplicationPersistenceExc
    *
    * @param cause the cause for the exception
    */
-  public InvalidFieldException(Throwable cause) {
+  public NotFoundException(Throwable cause) {
     super(cause);
   }
 }
