+++ replication-api/src/main/java/com/connexta/replication/api/data/TaskInfo.java
@@ -11,7 +11,7 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package com.connexta.replication.queue;
+package com.connexta.replication.api.data;
 
 import java.time.Instant;
 import java.util.Optional;
