+++ replication-api/src/main/java/com/connexta/replication/api/data/ddf/DDFMetadataInfo.java
@@ -11,9 +11,9 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package com.connexta.replication.queue.ddf;
+package com.connexta.replication.api.data.ddf;
 
-import com.connexta.replication.queue.MetadataInfo;
+import com.connexta.replication.api.data.MetadataInfo;
 
 /**
  * Extension to the {@link MetadataInfo} class which provides additional collected information that
