+++ replication-api/src/main/java/com/connexta/replication/api/persistence/DataManager.java
@@ -13,7 +13,9 @@
  */
 package com.connexta.replication.api.persistence;
 
+import com.connexta.replication.api.data.NotFoundException;
 import com.connexta.replication.api.data.Persistable;
+import com.connexta.replication.api.data.ReplicationPersistenceException;
 import java.util.stream.Stream;
 
 /**
@@ -27,10 +29,9 @@ public interface DataManager<T extends Persistable> {
    *
    * @param id The object id
    * @return the T object with the given id
-   * @throws com.connexta.ion.replication.api.NotFoundException if an object with the given id
-   *     cannot be found
-   * @throws com.connexta.ion.replication.api.ReplicationPersistenceException if any other error
-   *     occurs while trying to retrieve the object
+   * @throws NotFoundException if an object with the given id cannot be found
+   * @throws ReplicationPersistenceException if any other error occurs while trying to retrieve the
+   *     object
    * @throws IllegalStateException if multiple objects were found with the given id
    */
   T get(String id);
@@ -39,8 +40,7 @@ public interface DataManager<T extends Persistable> {
    * Gets all the currently saved T objects
    *
    * @return Stream of all T objects
-   * @throws com.connexta.ion.replication.api.ReplicationPersistenceException if an error occurs
-   *     while trying to retrieve the objects
+   * @throws ReplicationPersistenceException if an error occurs while trying to retrieve the objects
    */
   Stream<T> objects();
 
@@ -51,8 +51,7 @@ public interface DataManager<T extends Persistable> {
    *
    * @param object The T object to save or update
    * @throws IllegalArgumentException if the T implementation is not one that can be saved
-   * @throws com.connexta.ion.replication.api.ReplicationPersistenceException if an error occurs
-   *     while trying to save the object
+   * @throws ReplicationPersistenceException if an error occurs while trying to save the object
    */
   void save(T object);
 
@@ -60,10 +59,9 @@ public interface DataManager<T extends Persistable> {
    * Deletes a T object with the given id
    *
    * @param id The id of the object to be removed
-   * @throws com.connexta.ion.replication.api.NotFoundException if an object with the given id
-   *     cannot be found
-   * @throws com.connexta.ion.replication.api.ReplicationPersistenceException if any other error
-   *     occurs while trying to delete the object
+   * @throws NotFoundException if an object with the given id cannot be found
+   * @throws ReplicationPersistenceException if any other error occurs while trying to delete the
+   *     object
    */
   void remove(String id);
 }
